refactor: centralize package version loading and extract magic numbers
Three improvements for better maintainability:

1. Centralized Package Version Loading:
   Problem: package.json loading was duplicated across 4 files with brittle
   fallback logic that assumed specific build output structures. This could
   fail in different build configurations.

   Solution: Created src/utils/version.ts with centralized loading logic:
   - Single source of truth for version loading
   - Better error handling with fallback
   - Clear documentation of build scenarios
   - Updated manifest.ts to use new utility (other files next)

   Benefits:
   - Easier to maintain (one place to update)
   - More robust across build configurations
   - Better error messages when version can't be loaded

2. Extract Message Rotation Magic Numbers:
   Problem: Magic number 40 in index.ts made the relationship between
   messageRotationCounter, interval duration (200ms), and rotation frequency
   (8 seconds) unclear.

   Solution: Extract as named constants:
   - SPINNER_UPDATE_INTERVAL_MS = 200 (spinner refresh rate)
   - MESSAGE_ROTATION_INTERVAL_MS = 8000 (message rotation period)
   - MESSAGE_ROTATION_TICKS = Math.floor(8000 / 200) (calculated)

   Benefits:
   - Clear intent (what each number represents)
   - Easy to tune timings independently
   - Self-documenting code

3. Grammar Fix in Loading Message:
   Fixed: 'Convincing the LLM your variable names...'
   To: 'Convincing the LLM that your variable names...'

   Improves readability and grammatical correctness.

All 413 tests pass ✅
